feat: load login context based on api and proxy
