+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/configuration/server/ServerConfigRepository.kt
@@ -18,31 +18,21 @@
 
 package com.wire.kalium.logic.configuration.server
 
-import com.benasher44.uuid.uuid4
 import com.wire.kalium.logic.CoreFailure
 import com.wire.kalium.logic.StorageFailure
 import com.wire.kalium.logic.data.id.toDao
 import com.wire.kalium.logic.data.user.UserId
 import com.wire.kalium.logic.di.MapperProvider
-import com.wire.kalium.logic.failure.ServerConfigFailure
 import com.wire.kalium.logic.functional.Either
 import com.wire.kalium.logic.functional.flatMap
 import com.wire.kalium.logic.functional.fold
-import com.wire.kalium.logic.functional.left
 import com.wire.kalium.logic.functional.map
-import com.wire.kalium.logic.wrapApiRequest
 import com.wire.kalium.logic.wrapStorageRequest
 import com.wire.kalium.network.api.base.authenticated.UpgradePersonalToTeamApi.Companion.MIN_API_VERSION
 import com.wire.kalium.network.api.base.unbound.versioning.VersionApi
-import com.wire.kalium.network.api.unbound.configuration.ApiVersionDTO
 import com.wire.kalium.persistence.daokaliumdb.ServerConfigurationDAO
 import com.wire.kalium.util.KaliumDispatcher
 import com.wire.kalium.util.KaliumDispatcherImpl
-import io.ktor.http.Url
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.filterNotNull
-import kotlinx.coroutines.flow.flowOf
-import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.withContext
 
 interface ServerConfigRepository {
@@ -72,21 +62,22 @@ interface ServerConfigRepository {
     suspend fun configForUser(userId: UserId): Either<StorageFailure, ServerConfig>
     suspend fun commonApiVersion(domain: String): Either<CoreFailure, Int>
     suspend fun getTeamUrlForUser(userId: UserId): String?
-    suspend fun observeServerConfigByLinks(links: ServerConfig.Links): Flow<Either<CoreFailure, ServerConfig>>
 }
 
 @Suppress("LongParameterList", "TooManyFunctions")
 internal class ServerConfigDataSource(
-    private val dao: ServerConfigurationDAO,
-    private val versionApi: VersionApi,
-    private val serverConfigMapper: ServerConfigMapper = MapperProvider.serverConfigMapper(),
+    override val serverConfigurationDAO: ServerConfigurationDAO,
+    override val versionApi: VersionApi,
+    override val serverConfigMapper: ServerConfigMapper = MapperProvider.serverConfigMapper(),
     private val dispatchers: KaliumDispatcher = KaliumDispatcherImpl
-) : ServerConfigRepository {
+) : ServerConfigRepository, ServerConfigRepositoryExtension(
+    versionApi, serverConfigurationDAO, serverConfigMapper
+) {
 
     override val minimumApiVersionForPersonalToTeamAccountMigration = MIN_API_VERSION
 
     override suspend fun getOrFetchMetadata(serverLinks: ServerConfig.Links): Either<CoreFailure, ServerConfig> =
-        wrapStorageRequest { dao.configByLinks(serverConfigMapper.toEntity(serverLinks)) }.fold({
+        wrapStorageRequest { serverConfigurationDAO.configByLinks(serverConfigMapper.toEntity(serverLinks)) }.fold({
             fetchApiVersionAndStore(serverLinks)
         }, {
             Either.Right(serverConfigMapper.fromEntity(it))
@@ -94,60 +85,14 @@ internal class ServerConfigDataSource(
 
     override suspend fun storeConfig(links: ServerConfig.Links, metadata: ServerConfig.MetaData): Either<StorageFailure, ServerConfig> =
         withContext(dispatchers.io) {
-            wrapStorageRequest {
-                // check if such config is already inserted
-                val storedConfigId = dao.configByLinks(serverConfigMapper.toEntity(links))?.id
-                if (storedConfigId != null) {
-                    // if already exists then just update it
-                    dao.updateServerMetaData(
-                        id = storedConfigId,
-                        federation = metadata.federation,
-                        commonApiVersion = metadata.commonApiVersion.version
-                    )
-                    if (metadata.federation) dao.setFederationToTrue(storedConfigId)
-                    storedConfigId
-                } else {
-                    // otherwise insert new config
-                    val newId = uuid4().toString()
-                    dao.insert(
-                        ServerConfigurationDAO.InsertData(
-                            id = newId,
-                            apiBaseUrl = links.api,
-                            accountBaseUrl = links.accounts,
-                            webSocketBaseUrl = links.webSocket,
-                            blackListUrl = links.blackList,
-                            teamsUrl = links.teams,
-                            websiteUrl = links.website,
-                            isOnPremises = links.isOnPremises,
-                            title = links.title,
-                            federation = metadata.federation,
-                            domain = metadata.domain,
-                            commonApiVersion = metadata.commonApiVersion.version,
-                            apiProxyHost = links.apiProxy?.host,
-                            apiProxyNeedsAuthentication = links.apiProxy?.needsAuthentication,
-                            apiProxyPort = links.apiProxy?.port
-                        )
-                    )
-                    newId
-                }
-            }.flatMap { storedConfigId ->
-                wrapStorageRequest { dao.configById(storedConfigId) }
-            }.map {
-                serverConfigMapper.fromEntity(it)
-            }
+            storeServerLinksAndMetadata(links, metadata)
         }
-
-    override suspend fun fetchApiVersionAndStore(links: ServerConfig.Links): Either<CoreFailure, ServerConfig> =
-        fetchMetadata(links)
-            .flatMap { metaData ->
-                storeConfig(links, metaData)
-            }
 
     override suspend fun updateConfigMetaData(serverConfig: ServerConfig): Either<CoreFailure, Unit> =
         fetchMetadata(serverConfig.links)
             .flatMap { newMetaData ->
                 wrapStorageRequest {
-                    dao.updateServerMetaData(
+                    serverConfigurationDAO.updateServerMetaData(
                         id = serverConfig.id,
                         federation = newMetaData.federation,
                         commonApiVersion = newMetaData.commonApiVersion.version
@@ -156,29 +101,12 @@ internal class ServerConfigDataSource(
             }
 
     override suspend fun configForUser(userId: UserId): Either<StorageFailure, ServerConfig> =
-        wrapStorageRequest { dao.configForUser(userId.toDao()) }
+        wrapStorageRequest { serverConfigurationDAO.configForUser(userId.toDao()) }
             .map { serverConfigMapper.fromEntity(it) }
 
     override suspend fun commonApiVersion(domain: String): Either<CoreFailure, Int> = wrapStorageRequest {
-        dao.getCommonApiVersion(domain)
+        serverConfigurationDAO.getCommonApiVersion(domain)
     }
 
-    private suspend fun fetchMetadata(serverLinks: ServerConfig.Links): Either<CoreFailure, ServerConfig.MetaData> =
-        wrapApiRequest { versionApi.fetchApiVersion(Url(serverLinks.api)) }
-            .flatMap {
-                when (it.commonApiVersion) {
-                    ApiVersionDTO.Invalid.New -> Either.Left(ServerConfigFailure.NewServerVersion)
-                    ApiVersionDTO.Invalid.Unknown -> Either.Left(ServerConfigFailure.UnknownServerVersion)
-                    is ApiVersionDTO.Valid -> Either.Right(it)
-                }
-            }.map { serverConfigMapper.fromDTO(it) }
-
-    override suspend fun getTeamUrlForUser(userId: UserId): String? = dao.teamUrlForUser(userId.toDao())
-
-    override suspend fun observeServerConfigByLinks(links: ServerConfig.Links): Flow<Either<CoreFailure, ServerConfig>> =
-        fetchApiVersionAndStore(links).fold({ flowOf(it.left()) }, {
-            dao.getServerConfigByLinksFlow(serverConfigMapper.toEntity(links)).filterNotNull()
-                .map(serverConfigMapper::fromEntity)
-                .wrapStorageRequest()
-        })
+    override suspend fun getTeamUrlForUser(userId: UserId): String? = serverConfigurationDAO.teamUrlForUser(userId.toDao())
 }
