+++ network/src/commonMain/kotlin/com/wire/kalium/network/api/base/unbound/versioning/VersionApi.kt
@@ -21,6 +21,7 @@ package com.wire.kalium.network.api.base.unbound.versioning
 import com.wire.kalium.network.BackendMetaDataUtil
 import com.wire.kalium.network.BackendMetaDataUtilImpl
 import com.wire.kalium.network.UnauthenticatedNetworkClient
+import com.wire.kalium.network.UnboundNetworkClient
 import com.wire.kalium.network.api.unbound.configuration.ServerConfigDTO
 import com.wire.kalium.network.api.unbound.versioning.VersionInfoDTO
 import com.wire.kalium.network.utils.NetworkResponse
@@ -41,6 +42,11 @@ class VersionApiImpl internal constructor(
     private val developmentApiEnabled: Boolean,
     private val util: BackendMetaDataUtil = BackendMetaDataUtilImpl
 ) : VersionApi {
+
+    internal constructor(unboundNetworkClient: UnboundNetworkClient, developmentApiEnabled: Boolean) : this(
+        httpClient = unboundNetworkClient.httpClient,
+        developmentApiEnabled = developmentApiEnabled
+    )
 
     internal constructor(unauthenticatedNetworkClient: UnauthenticatedNetworkClient, developmentApiEnabled: Boolean) : this(
         unauthenticatedNetworkClient.httpClient,
